+++ semver-gradle-plugin/main/kotlin/com/javiersc/semver/gradle/plugin/tasks/PushSemverTagTask.kt
@@ -1,7 +1,7 @@
 package com.javiersc.semver.gradle.plugin.tasks
 
+import com.javiersc.semver.gradle.plugin.internal.projectTagPrefix
 import com.javiersc.semver.gradle.plugin.internal.tagPrefixProperty
-import com.javiersc.semver.gradle.plugin.semverExtension
 import com.javiersc.semver.gradle.plugin.services.GitBuildService
 import org.gradle.api.DefaultTask
 import org.gradle.api.Project
@@ -38,7 +38,7 @@ public abstract class PushSemverTagTask : DefaultTask() {
         internal fun register(project: Project, gitTagBuildService: Provider<GitBuildService>) {
             project.tasks.register<PushSemverTagTask>(taskName).configure {
                 this.tagPrefixProperty.set(project.tagPrefixProperty)
-                this.projectTagPrefix.set(project.semverExtension.tagPrefix)
+                this.projectTagPrefix.set(project.projectTagPrefix)
                 this.version.set(project.version.toString())
                 this.gitTagBuildService.set(gitTagBuildService)
                 this.usesService(gitTagBuildService)